Improve peak detection and remove showing raw data in visualization
base the peak distance detection on the time between peaks, with some variety to also handle noiser data, like that recorded with mic
showing the raw data in visualization gave no purpose but to use even more CPU and memory and slow it down. Use something like Audacity to see the raw data.